requirements.txt
ultralytics>=8.0.0
opencv-python>=4.5.0
numpy>=1.20.0
PyYAML>=6.0
Pillow>=9.0.0
torch>=2.0.0
torchvision>=0.10.0
tqdm>=4.62.0
pytesseract>=0.3.8 